annotation_gui_gcp/requirements.txt
flask
python-magic
